requirements.txt
python-binance==1.0.15
tensorflow-cpu==2.8.0
pandas==1.4.1
numpy==1.22.3
gcsfs==2022.3.0
google-cloud-storage==2.3.0

# s3fs==2022.3.0
# google-cloud==0.34.0
# scikit-learn==1.0.2
# matplotlib==3.5.1
# TA-Lib==0.4.24